Move line chart dialog.
